+++ CHANGELOG.md
@@ -5,4 +5,5 @@
 - [Refactoring:remove invalid @AutoConfigureAfter and @AutoConfigureBefore from discovery client automatic configuration.](https://github.com/Tencent/spring-cloud-tencent/pull/1116)
 - [fix:fix feign url bug when using sleuth.](https://github.com/Tencent/spring-cloud-tencent/pull/1120)
 - [refactor:optimize the order and condition matching of service registration automatic configuration.](https://github.com/Tencent/spring-cloud-tencent/pull/1131)
+- [feat:support service contract reporting.](https://github.com/Tencent/spring-cloud-tencent/pull/1141)
 - [feat: support log path configuration parameters,](https://github.com/Tencent/spring-cloud-tencent/pull/1142)